Refactor table get records method
